Split sax code into separate file from benchmark
